DSND-1566: Simplifying charges processing * Tested locally * removed ChargesDataApiService as unnecessary
